+++ xivo_confd/plugins/register_iax/service.py
@@ -3,13 +3,14 @@
 # SPDX-License-Identifier: GPL-3.0+
 
 from xivo_confd.helpers.resource import CRUDService
-from xivo_dao.resources.register_sip import dao as register_sip_dao
 
-from .validator import build_validator
+from xivo_dao.resources.register_iax import dao as register_iax_dao
+
 from .notifier import build_notifier
+from .validator import build_validator
 
 
 def build_service():
-    return CRUDService(register_sip_dao,
+    return CRUDService(register_iax_dao,
                        build_validator(),
                        build_notifier())